Make rabbit_fifo_dlx not depend on rabbit_fifo
so that we won't have to copy rabbit_fifo_dlx.erl file for new
ra machine versions in the future.

Consequences:
1. rabbit_fifo_dlx maintains its own ra_indexes.
2. rabbit_fifo_dlx does not modify rabbit_fifo's messages_total.
   Instead, rabbit_fifo asks rabbit_fifo_dlx for its total messages.

This commit still maintains correct dehydration + snapshotting for
in-memory messages. However, after this commit, messages in
rabbit_fifo_dlx do not count as in-memory messages anymore even
when they are in-memory. That's okay given that we remove in-memory
messages shortly via
https://github.com/rabbitmq/rabbitmq-server/issues/3898